i965: Update 3DSTATE_{DEPTH,STENCIL,...}_BUFFER and such for Broadwell.
(This patch was written before Paul's refactor to reuse
 brw_emit_depthbuffer and just implement a gen7_emit_depth_stencil_hiz
 function.  We should probably rearchitect it to fit that new model, as
 it's much better.  --Ken)

(The rest of this commit message was written by Chad.)

The depth buffer packet emitted by this patchis incomplete.

We must use the packet's level/layer fields to specify the miptree slice
into which to render. This patch emits 0 there, and hence all rendering
occurs to the miptree's base slice. (In gen5 through gen7, we used the
intra-tile x/y offsets to specify the miptree slice. But gen8 removed
those fields.)

However, the incomplete implementation does show some signs of life.  It
fixes the page fault in piglit:fbo-depth. (The test's segfault was
likely due to the high 16 bits of the surface base address being
invalid).

v2: [chadv]
  - Comment in patch that depth buffer packet needs level/layer
    fields.
  - Remove comments specific to gen7.
  - Expand commit message.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>